rename: 컨벤션에 맞추어 폴더구조 변경 (#111)
